frontend setup

+++ backend/controllers/product.controller.js
@@ -39,19 +39,24 @@ export const updateProduct = async (req , res) => {
         res.status(201).json({success:true , data:updatedProduct});
     }
     catch(error){
-        req.status(500).json({success:false, message:"Server error"});
+        req.status(500).json({success:false, message:"Internal server error."});
     } 
 }; 
 
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params  ;
+    
+    if(!product.name || !product.price || !product.image) {
+        return res.status(400).json({ success: false , message: "Fillup all fields" }) ;
+    }
+    
     try{ 
         await Product.findByIdAndDelete(id) ;
         res.status(200).json({ success: true, message: "Product deleted successfully" }) ; 
     }
     catch(error){
         console.error("Error in delete product: ", error) ;
-        res.status(404).json({ success: false, message: "Product not found" }) ;
+        res.status(500).json({ success: false, message: "Internal server Error." }) ;
     }
 };
